Previous brute force solution actually O(n^2) due to string concatenation
This solution is O(n) in cases where the compressed string is shorter
than the original. Worst case would still be O(n^2). However, we can
re-purpose the code that performs the compression to count the final
length of the resultant string and create a StringBuilder of that size
to maintain a O(n) runtime for all cases in exchange for ~2x the amount
of code.